+++ test/lib/nibble/upgrade_test.rb
@@ -4,7 +4,7 @@ class Nibble::UpgradeTest < ActiveSupport::TestCase
   setup do
     @steps = []
     @themes = Pathname(Dir.mktmpdir("nibble-upgrade"))
-    FileUtils.cp_r(Rails.root.join("themes/crumbs"), @themes.join("crumbs"))
+    FileUtils.cp_r(Nibble.core_root.join("themes/crumbs"), @themes.join("crumbs"))
     Nibble.config = Nibble::Config.new(Nibble.config_values.merge("theme" => "crumbs"),
       themes_path: @themes, site_schema_path: @themes.join("no_site_schema"))
     Nibble.reset_schema!